+++ contracts/cashier/token/IUploadToken.sol
@@ -3,7 +3,7 @@ pragma solidity >=0.8.0 <0.9.0;
 import "../../token/ISafeERC20.sol";
 
 interface IUploadToken is ISafeERC20 {
-    function mintForMarket(uint256 amount) external;
+    function mintForMarket(uint amount) external;
 
-    function consume(address consumer, uint256 amount) external;
+    function consume(address consumer, uint amount) external;
 }
