+++ contracts/interfaces/IReward.sol
@@ -3,19 +3,9 @@
 pragma solidity >=0.8.0 <0.9.0;
 
 interface IReward {
-    event DistributeReward(
-        uint256 indexed pricingIndex,
-        address indexed beneficiary,
-        uint256 amount
-    );
+    event DistributeReward(uint indexed pricingIndex, address indexed beneficiary, uint amount);
 
-    function fillReward(uint256 beforeLength, uint256 rewardSectors)
-        external
-        payable;
+    function fillReward(uint beforeLength, uint rewardSectors) external payable;
 
-    function claimMineReward(
-        uint256 pricingIndex,
-        address payable beneficiary,
-        bytes32 minerId
-    ) external;
+    function claimMineReward(uint pricingIndex, address payable beneficiary, bytes32 minerId) external;
 }
