+++ contracts/test/ExponentTest.sol
@@ -4,11 +4,11 @@ pragma solidity >=0.8.0 <0.9.0;
 import "../utils/Exponent.sol";
 
 contract ExponentTest {
-    function powTwo(uint256 exponentX64) public pure returns (uint256) {
+    function powTwo(uint exponentX64) public pure returns (uint) {
         return Exponential.powTwo64X96(exponentX64);
     }
 
-    function powHalf(uint256 exponentX64) public pure returns (uint256) {
+    function powHalf(uint exponentX64) public pure returns (uint) {
         return Exponential.powHalf64X96(exponentX64);
     }
 }
