+++ contracts/token/MockHackToken.sol
@@ -12,11 +12,7 @@ contract MockHackToken is ERC20, ISafeERC20 {
         _mint(msg.sender, 1000);
     }
 
-    function _afterTokenTransfer(
-        address,
-        address to,
-        uint256 amount
-    ) internal override {
+    function _afterTokenTransfer(address, address to, uint amount) internal override {
         if (to == address(0)) {
             return;
         }
