+++ contracts/uploadMarket/ICashier.sol
@@ -2,10 +2,9 @@
 pragma solidity >=0.8.0 <0.9.0;
 
 interface ICashier {
-    function chargeFee(uint256 uploadSectors, uint256 paddingSectors) external;
+    function chargeFee(uint uploadSectors, uint paddingSectors) external;
 
-    function claimMineReward(uint256 pricingIndex, address beneficiary)
-        external;
+    function claimMineReward(uint pricingIndex, address beneficiary) external;
 
-    function getFlowLength() external view returns (uint256);
+    function getFlowLength() external view returns (uint);
 }
